requirements.txt
Django
djangorestframework
Pillow
psycopg2-binary
django-filter
pyjwt
drf-yasg==1.20
python-decouple